requirements.txt
# Azure Key Vault dependencies
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0

# YAML processing
PyYAML>=6.0

# Additional utilities
requests>=2.28.0 